+++ TODO.R
@@ -20,7 +20,6 @@
 # TODO [delete] generated files
 # TODO [proj_file] to accesss data (return the path in this case?)
 # TODO [check_referenced_files] doesn't check for {.file R/file.R}
-# TODO explain rationale behind `work_only`. Suggest to transform to TODO if this item is no longer critical. `work_only` goal is to show you exactly where you need to do work
 # TODO browse_pkg should open by default if no vignettes are found, because there is not much to do in the R-session.
 # TODO exclude _files from `proj_list()`
 # TODO [outline] Show function call if exported + not internal + bonus if has family tag! rstudio/rstudio#14766