+++ tests/testthat/test-outline-criteria.R
@@ -25,16 +25,12 @@ test_that("o_is_todo_fixme() works", {
   # avoid finding comments.
   expect_false(o_is_todo_fixme("# another TODO item"))
   expect_false(o_is_todo_fixme("#' TODO, WORK, FIXME)", is_roxygen_comment = T))
-
 })
 
-test_that("o_is_work_item() works", {
-  expect_true(o_is_work_item("# WORK this needs to be done."))
-})
-
-test_that("o_is_test_that() works", {
-  expect_true(o_is_test_that('test_that("Serious things are happening", {'))
-  expect_false(o_is_test_that('test_that("", {'))
+test_that("o_is_test_name() works", {
+  expect_true(o_is_test_name('test_that("Serious things are happening", {'))
+  expect_true(o_is_test_name('describe("This is happening", {'))
+  expect_false(o_is_test_name('test_that("", {'))
 })
 
 test_that("o_is_generic_test() works", {